pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "remembrall-starter"
version = "0.1.0"
description = "Starter kit for Remembrall-style on-device continuous learning."
readme = "README.md"
requires-python = ">=3.9"
authors = [{name="Remembrall Authors"}]
license = {text = "MIT"}
dependencies = [
    "numpy>=1.22",
    "scipy>=1.10",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
torch = ["torch>=2.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]
